EZP-32271: Split commands responsibility
